+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 #### Dependency Upgrade
 
 #### New Features
+* Added DSL support for `admissionregistration.k8s.io/v1beta1` resources
 * Add support for Namespaced SharedInformers, fixed probelms with OperationContext argument
 
 ### 4.10.1 (2020-05-06)